ensemble refactoring
